# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the Incident API server.",
        "title": "Incident API",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "incident.deltalab.dev",
    "basePath": "/v0",
    "paths": {
        "/incident/help": {
            "post": {
                "description": "Help",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Incident"
                ],
                "summary": "Help",
                "parameters": [
                    {
                        "description": "Help request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/incident.HelpRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/incident/help/query": {
            "post": {
                "description": "Query help",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Incident"
                ],
                "summary": "Query help",
                "parameters": [
                    {
                        "description": "Query help request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/incident.QueryHelpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/incident.QueryHelpResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/incident/image/{incident_id}": {
            "get": {
                "description": "Query incident image",
                "tags": [
                    "Incident"
                ],
                "summary": "Query incident image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Incident ID",
                        "name": "incident_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/incident/query": {
            "post": {
                "description": "Query incidents",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Incident"
                ],
                "summary": "Query incidents",
                "parameters": [
                    {
                        "description": "Query request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/incident.QueryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/incident.QueryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/incident/report": {
            "post": {
                "description": "Report an incident",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Incident"
                ],
                "summary": "Report an incident",
                "parameters": [
                    {
                        "description": "Report request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/incident.ReportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/incident.ReportResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an incident report",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Incident"
                ],
                "summary": "Delete an incident report",
                "parameters": [
                    {
                        "description": "Delete report request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/incident.DeleteReportRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "incident.DeleteReportRequest": {
            "type": "object",
            "properties": {
                "incident_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "user_key": {
                    "type": "string"
                }
            }
        },
        "incident.HelpRequest": {
            "type": "object",
            "properties": {
                "blood_pressure": {
                    "type": "integer"
                },
                "body_temperature": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "heart_rate": {
                    "type": "integer"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "incident.QueryHelpRequest": {
            "type": "object",
            "properties": {
                "help_id": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "radius": {
                    "description": "Radius in meters",
                    "type": "integer"
                }
            }
        },
        "incident.QueryHelpResponse": {
            "type": "object",
            "properties": {
                "blood_pressure": {
                    "type": "integer"
                },
                "body_temperature": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "heart_rate": {
                    "type": "integer"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "reporter": {
                    "type": "string"
                }
            }
        },
        "incident.QueryRequest": {
            "type": "object",
            "properties": {
                "incident_id": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "radius": {
                    "description": "Radius in meters",
                    "type": "integer"
                }
            }
        },
        "incident.QueryResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "incident_id": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "reporter": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "incident.ReportRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "image": {
                    "description": "Base64 encoded jpeg image",
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "type": {
                    "type": "string"
                },
                "user_id": {
                    "description": "user id in uuid format",
                    "type": "string"
                },
                "user_key": {
                    "description": "32 byte ascii string",
                    "type": "string"
                }
            }
        },
        "incident.ReportResponse": {
            "type": "object",
            "properties": {
                "incident_id": {
                    "type": "string"
                }
            }
        }
    }
}